# openapi/orbithr-v1.yaml
openapi: 3.1.0
info:
  title: OrbitHR API
  version: 1.0.0
  description: Authenticated, tenant-isolated API for OrbitHR web and Android clients.
servers:
  - url: http://localhost:3001/api/v1
security:
  - bearerAuth: []
paths:
...
  /organization:
    get:
      operationId: getOrganizationStructure
      responses:
        '200': {description: Tenant-scoped branches, locations, departments, teams, designations, cost centres, and grades}
        '403': {$ref: '#/components/responses/ErrorResponse'}
...
  /rbac:
    get:
      operationId: getAccessConfiguration
      responses:
        '200': {description: Tenant permissions, roles, and scoped grants}
        '403': {$ref: '#/components/responses/ErrorResponse'}
...
  /rbac/grants/{id}:
    delete:
      operationId: revokeAccessGrant
      parameters:
        - {in: path, name: id, required: true, schema: {type: string, format: uuid}}
      responses:
        '200': {description: Access grant revoked}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: {type: string}
            message: {type: string}
            fieldErrors: {type: object}
        meta:
          type: object
          properties:
            requestId: {type: string, format: uuid}
  responses:
...
    ErrorResponse:
      description: Standard API error
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}